Moved get_email method to base_form module
